refactor: 폴더 구조를 shared/types로 변경하고 import 경로 수정 - 루트 types.ts를 shared/types/ 폴더로 이동 - 계층 구조에 맞게 타입 참조 경로 수정
