+++ shared/types/index.ts
@@ -1,0 +1,11 @@
+// 공통 타입 모음 인덱스
+// shared/types/index.ts
+
+// 공통 응답 및 요청 인터페이스
+export * from './common';
+
+// 인증 관련 타입
+export * from './auth';
+
+// API 관련 타입
+export * from './api';
